split up keeper/validator.go
